refactor: move interfaces to the types.ts
